pyproject.toml
[project]
name = "homework-5-python-django-rest-framework"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksim Lakovich",email = "maks.lakovich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "mypy-extensions (>=1.0.0,<2.0.0)",
    "django-stubs (>=5.1.3,<6.0.0)",
    "djangorestframework-stubs (>=3.15.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
django-phonenumber-field = "^8.0.0"
django-stubs = "^5.1.3"

